fixing bugs
