+++ src/components/Play/PlaySetupPage.jsx
@@ -1,4 +1,4 @@
-import React, { useContext, useRef, useEffect } from "react";
+import React, { useContext, useRef } from "react";
 import { AvatarCtx } from "index";
 import { Redirect } from "react-router";
 import { Image, Card, Button } from "antd";
@@ -6,9 +6,9 @@ import { RightOutlined, LeftOutlined } from "@ant-design/icons";
 import { Link } from "react-router-dom";
 import { BtnPrimary } from "../../GlobalStyles";
 import { NFTImg, BreakFlexDiv, brightFontCol } from "../../GlobalStyles";
-import Webcam from "react-webcam";
 import SelectWebcam from "components/Webcam/SelectWebcam";
 import { WebcamCtx } from "index";
+import PoseDetWebcam from "components/Webcam/PoseDetWebcam";
 
 const styles = {
     titleText: {
@@ -33,74 +33,7 @@ const styles = {
 
 const PlaySetupPage = () => {
     const [avatar] = useContext(AvatarCtx);
-    const [webcamId, setWebcamId] = useContext(WebcamCtx);
-    const webcamRef = useRef(null);
-    const canvasRef = useRef(null);
-
-    console.log('webcamId', webcamId);
-
-    const runCoco = async () => {
-        requestAnimationFrame(() => {
-            runCoco()
-        })
-        // 3. TODO - Load network 
-        // e.g. const net = await cocossd.load();
-
-        //  Loop and detect hands
-        // setInterval(() => {
-        //     detect();
-        // }, 10);
-        // detect();
-
-        detect()
-    };
-
-    const detect = async () => {
-        // Check data is available
-        if (
-            typeof webcamRef.current !== "undefined" &&
-            webcamRef.current !== null &&
-            webcamRef.current.video.readyState === 4
-        ) {
-            // Get Video Properties
-
-            const video = webcamRef.current.video;
-            const videoWidth = webcamRef.current.video.videoWidth;
-            const videoHeight = webcamRef.current.video.videoHeight;
-
-            // const video = webcamRef.current.video;
-            // const videoWidth = webcamRef.current.video.videoWidth;
-            // const videoHeight = webcamRef.current.video.videoHeight;
-
-            // Set video width
-            // webcamRef.current.video.width = videoWidth;
-            // webcamRef.current.video.height = videoHeight;
-
-            // Set canvas height and width
-            canvasRef.current.width = videoWidth;
-            canvasRef.current.height = videoHeight;
-
-            // 4. TODO - Make Detections
-            // e.g. const obj = await net.detect(video);
-
-            // Draw mesh
-            const ctx = canvasRef.current.getContext("2d");
-            // ctx.drawImage(
-            //     video, 0, 0, videoWidth, videoHeight
-            // );
-
-            ctx.fillStyle = '#FF0300';
-            ctx.strokeStyle = '#FF0300';
-            ctx.beginPath();
-            ctx.arc(95, 50, 40, 0, 2 * Math.PI);
-            ctx.stroke();
-            ctx.fill();
-            // 5. TODO - Update drawing utility
-            // drawSomething(obj, ctx)  
-        }
-    };
-
-    useEffect(() => { runCoco() }, []);
+    const {webcamId, setWebcamId} = useContext(WebcamCtx);
 
     if (!avatar) {
         return <Redirect to="/avatars" />;
@@ -109,46 +42,11 @@ const PlaySetupPage = () => {
     return (<>
 
         {/*  */}
-        <>
-            <Webcam
-                audio={false}
-                videoConstraints={{ deviceId: webcamId }}
-                mirrored={true}
-                className={"webcam"}
-                ref={webcamRef}
-                muted={true}
-                style={{
-                    objectFit: "cover",
-                    borderRadius: "1rem",
-                    boxShadow: "0 0 10px 2px #202020",
-                    position: "absolute",
-                    height: "auto",
-                    zindex: 9,
-                    // param
-                    top: "10%",
-                    left: "51%",
-                    width: "20%",
-                }}
-            />
-            <canvas
-                ref={canvasRef}
-                className={"webcam-canvas"}
-                style={{
-                    objectFit: "cover",
-                    borderRadius: "1rem",
-                    position: "absolute",
-                    height: "auto",
-                    zindex: 8,
-                    // debug
-                    border: "5px solid red",
-                    // param
-                    top: "10%",
-                    left: "51%",
-                    width: "20%",
-                }}
-            />
-
-        </>
+        <PoseDetWebcam
+            camWidth={"20%"}
+            camTop={"10%"}
+            camLeft={"51%"}
+        />
 
         {/* main cards */}
         <div style={{
@@ -219,7 +117,7 @@ const PlaySetupPage = () => {
                         paddingBottom: "1rem",
                         textAlign: "center",
                     }}>
-                        <SelectWebcam webcamRef={webcamRef} width={"15rem"} />
+                        <SelectWebcam width={"15rem"} />
                     </div>
                     <div style={{
                         ...BreakFlexDiv,
